[emontx] Add typing, register pattern, uint16_t buffer_pos, DRY triggers
- Add return types to final_validate, to_code, action_to_code
- Add ConfigType hints to all config parameters
- Replace full_config scanning with register_sensor pattern
- Change buffer_pos_ from size_t to uint16_t
- Remove dead .get() fallback for rx_buffer_size (always has default)
- DRY on_json/on_data trigger blocks into a loop

+++ esphome/components/emontx/emontx.h
@@ -58,7 +58,7 @@ class EmonTx : public Component, public uart::UARTDevice {
 #endif
   LazyCallbackManager<void(JsonObject, StringRef)> json_callbacks_;
   LazyCallbackManager<void(StringRef)> data_callbacks_;
-  size_t buffer_pos_{0};
+  uint16_t buffer_pos_{0};
   std::array<char, MAX_LINE_LENGTH + 1> buffer_{};
 };
